requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
pandas
numpy
scikit-learn
xgboost
torch
transformers
requests
alpha_vantage
beautifulsoup4
lxml
feedparser
matplotlib
